+++ mobile/cmp/navigator/NavigatorModel.js
@@ -31,6 +31,9 @@ export class NavigatorModel extends HoistModel {
     /** @member {boolean} */
     swipeToGoBack;
 
+    /** @member {boolean} */
+    pullDownToRefresh;
+
     /** @member {RenderMode} */
     renderMode;
 
@@ -57,6 +60,7 @@ export class NavigatorModel extends HoistModel {
      * @param {boolean} [track] - True to enable activity tracking of page views (default false).
      *      Viewing of each page will be tracked with the `oncePerSession` flag, to avoid duplication.
      * @param {boolean} [swipeToGoBack] - True to enable 'swipe to go back' functionality.
+     * @param {boolean} [pullDownToRefresh] - True to enable 'pull down to refresh' functionality.
      * @param {RenderMode} [renderMode] - strategy for rendering pages. Can be set per-page
      *      via `PageModel.renderMode`. See enum for description of supported modes.
      * @param {RefreshMode} [refreshMode] - strategy for refreshing pages. Can be set per-page
@@ -66,6 +70,7 @@ export class NavigatorModel extends HoistModel {
         pages,
         track = false,
         swipeToGoBack = true,
+        pullDownToRefresh = true,
         renderMode = RenderMode.LAZY,
         refreshMode = RefreshMode.ON_SHOW_LAZY
     }) {
@@ -79,6 +84,7 @@ export class NavigatorModel extends HoistModel {
         this.pages = pages;
         this.track = track;
         this.swipeToGoBack = swipeToGoBack;
+        this.pullDownToRefresh = pullDownToRefresh;
         this.renderMode = renderMode;
         this.refreshMode = refreshMode;