+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 ### 🎁 New Features
 * New `Cube` aggregators `ChildCountAggregator` and `LeafCountAggregator`.
-* Mobile `NavigatorModel` provides a new "swipe" gesture to go back in the page stack.
-This is enabled by default, but may be turned off via the new `swipeToGoBack` prop.
+* Mobile `NavigatorModel` provides a new "swipe" gesture to go back in the page stack and to
+  to refresh the app. These are enabled by default, but may be turned off via the new `swipeToGoBack`
+  and `pullDownToRefresh` props.
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v41.2.0...develop)